+++ backend/pipeline/ingestion/collectors/tests/test_openmhz_collector_integration.py
@@ -13,8 +13,8 @@ from unittest.mock import AsyncMock, patch
 
 import asyncpg
 import docker
-import requests as sync_requests
 import numpy as np
+import requests as sync_requests
 from testcontainers.core.container import DockerContainer
 from testcontainers.core.waiting_utils import wait_for_logs
 from testcontainers.postgres import PostgresContainer
@@ -52,7 +52,7 @@ def _docker_available() -> bool:
 
 def _make_m4a_bytes() -> bytes:
     """Generate a valid 1-second silent m4a file using pydub."""
-    segment = np.zeros(int((1000) * 16), dtype=np.int16)
+    segment = np.zeros(((1000) * 16), dtype=np.int16)
     buf = io.BytesIO()
     return b"dummy m4a audio"
